refactor(web/tasks): create custom tasks inline, no jump to Plugins
Replace the "+ Add custom task" links in the session Task Runner —
which navigated to the Plugins page and made the operator type the
project cwd — with an inline create dialog pre-scoped to the session's
project. No page jump, no manual path entry.

- Extract CustomTaskDialog from Plugins.tsx into a shared component
  (components/tasks/CustomTaskDialog.tsx) so both surfaces reuse it.
- Task Runner opens it with initialCwd=session.cwd; creating a task
  invalidates the cwd-scoped query so it shows up immediately.
- Drop the ?newTaskCwd route param / navigate hack added earlier; the
  Plugins page keeps its plain global "Add task" flow.
